requirement.txt
fastapi
uvicorn
pymongo[srv]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
ytmusicapi
yt-dlp
dnspython
cachetools
requests